backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
asgiref==3.7.2 